Move get_context_data to mixin.py
We have need for a view which is external to this app to use the same
logic for filling the context.  To obtain DRY, we pop out the logic from
the X509ListView and place it in a mixin.  We then mix said mixin back
in (we just don't say it three times fast).
